+++ packages/nestjs-endpoints/src/setup-openapi.ts
@@ -3,7 +3,7 @@ import path from 'node:path';
 import { Writable } from 'node:stream';
 import { INestApplication } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { patchNestJsSwagger } from 'nestjs-zod';
+import { settings } from './consts';
 import { getCallsiteFile } from './helpers';
 
 async function readPkgJson() {
@@ -31,13 +31,19 @@ export async function setupOpenAPI(
     outputFile?: string | Writable;
   },
 ) {
-  patchNestJsSwagger();
   const builder = new DocumentBuilder();
   if (options?.configure) {
     options.configure(builder);
   }
   const config = builder.build();
   const document = SwaggerModule.createDocument(app, config);
+  document.components = {
+    ...document.components,
+    schemas: {
+      ...document.components?.schemas,
+      ...settings.openapi.components.schemas,
+    },
+  };
   Object.assign(document, {
     info: {
       ...document.info,