+++ packages/nestjs-endpoints/src/z.ts
@@ -1,0 +1,7 @@
+import 'zod-openapi/extend';
+import { z } from 'zod';
+import { extendZodWithOpenApi } from 'zod-openapi';
+
+extendZodWithOpenApi(z);
+
+export { z };