+++ packages/nestjs-endpoints/vitest.config.ts
@@ -1,0 +1,10 @@
+import path from 'path';
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    globals: true,
+    root: path.join(__dirname, '../..'),
+    include: [path.join(__dirname, '**/*.spec.ts')],
+  },
+});